requirements.txt
streamlit==1.28.0
pandas==2.1.2
beautifulsoup4==4.12.2
matplotlib==3.8.1
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.0.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.1
fpdf2==2.8.4
rapidfuzz==2.9.0